+++ app/notifiers/company_notifier.rb
@@ -6,35 +6,16 @@ class CompanyNotifier < AbstractNotifier::Base
   def insights_report = report
   def repository_insights_report = report
   def long_time_review_report = report
+  def no_new_pulls_report = report
 
   private
 
   def report
     notification(
       **company_payload.call(
-        type: source,
-        company: company,
         notification_name: notification_name,
-        notifications: notifications
+        notification: params[:notification]
       )
     )
-  end
-
-  def company = params[:company]
-
-  def notifications
-    company
-      .notifications
-      .joins(:webhook)
-      .where(notification_type: notification_type)
-      .where(webhooks: { source: source })
-  end
-
-  def notification_type
-    case notification_name
-    when :insights_report then Notification::INSIGHTS_DATA
-    when :repository_insights_report then Notification::REPOSITORY_INSIGHTS_DATA
-    when :long_time_review_report then Notification::LONG_TIME_REVIEW_DATA
-    end
   end
 end