+++ app/notifiers/company_webhook_notifier.rb
@@ -2,13 +2,9 @@
 
 class CompanyWebhookNotifier < CompanyNotifier
   self.driver = proc do |data|
-    data[:paths].each do |url|
-      Webhooks::Client.new(url: url.origin).send_message(
-        path: url.path,
-        body: data[:body]
-      )
-    end
+    Webhooks::Client.new(url: data[:path].origin).send_message(
+      path: data[:path].path,
+      body: data[:body]
+    )
   end
-
-  def source = Webhook::CUSTOM
 end