+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - time_since_last_open_pull_seconds insight
+- notifications for long time_since_last_open_pull_seconds
 
 ### Fixed
 - multiple notifications sending